Refactored User Story 14
- Renamed `#validate_dates` and `#check_bookings` helper methods for
  clarity.
- Refactored `#validate` to simplify date validation method calls.
- Extracted all helper methods from `app/app.rb` to separate
  `helpers.rb` file.
- Updated Request model to make date fields required.
- Updated `views/spaces.erb` and `views/view_space.erb` to make date
  inputs required for form submission.
