+++ app/models/request.rb
@@ -2,8 +2,8 @@ class Request
   include DataMapper::Resource
 
   property :id, Serial
-  property :start_date, Date
-  property :end_date, Date
+  property :start_date, Date, required: true
+  property :end_date, Date, required: true
   property :status, String
 
   belongs_to :user
