+++ app/views/spaces.erb
@@ -10,17 +10,17 @@
     <form action="/spaces/filter" method="post">
         <label>
             Available from
-            <input type="date" name="date_from" value="<%= @date_from %>">
+            <input type="date" name="date_from" value="<%= @date_from %>" required />
         </label>
         <label>
             Available to
-            <input type="date" name="date_to" value="<%= @date_to %>">
+            <input type="date" name="date_to" value="<%= @date_to %>" required />
         </label>
-        <input type="submit" value="Filter Spaces">
+        <input type="submit" value="Filter Spaces" />
     </form>
     <form action="/spaces/filter-reset" method="post">
-        <input type="hidden" name="_method" value="delete">
-        <input type="submit" value="Reset">
+        <input type="hidden" name="_method" value="delete" />
+        <input type="submit" value="Reset" />
     </form>
     <ul><% @spaces.each do |space| %>
         <li>
@@ -32,8 +32,8 @@
                 <p>Price: £<%= sprintf("%0.02f", space.price) %></p>
                 <% if space.user_id == @user.id %>
                     <form action="/spaces/update" method="POST">
-                        <input type="hidden" name="space_id" value="<%= space.id %>"></input>
-                        <input type="submit" value="Update Space">
+                        <input type="hidden" name="space_id" value="<%= space.id %>" />
+                        <input type="submit" value="Update Space" />
                     </form>
                 <% end %>
             </article>
@@ -43,5 +43,5 @@
 <% end %>
 
 <form action="/spaces/new" method="get">
-    <input type="submit" value="List a Space">
+    <input type="submit" value="List a Space" />
 </form>
